requirements.txt
spacy
odoorpc
openai-whisper
sounddevice
numpy